feat(vulkan): improve Vulkan command buffer readiness checks and error handling
- Added checks to skip rendering operations if Vulkan command buffer is not ready or fully initialized, preventing potential crashes.
- Enhanced validation of Vulkan device operations to catch device loss errors early and provide clearer diagnostics.
- Implemented comprehensive tests for Vulkan initialization, including memory allocation and image creation, to ensure stability.
- Updated error handling in Vulkan error checking macro to provide more descriptive messages for device-related issues.

+++ src/common/vm.c
@@ -2260,7 +2260,7 @@ static qboolean VM_ValidateVMState( vm_t *vm ) {
 
 	// Validate function pointers if they exist (cast to void* for validation)
 	if ( vm->destroy ) {
-		if ( !VM_ValidatePointer( (void *)vm->destroy, "vm_destroy" ) ) {
+		if ( !VM_ValidatePointer( (void *)(uintptr_t)vm->destroy, "vm_destroy" ) ) {
 			Com_Printf( "VM_ValidateVMState: VM destroy function pointer appears corrupted\n" );
 			return qfalse;
 		}
